requirements.txt
Pillow==8.3.2
Django==3.2.7
djangorestframework==3.12.4
django_q==1.3.9
